+++ ExcellaPracticeProject/ExcellaProject/src/main/java/com/excellaproject/repositories/Form_CRepository.java
@@ -8,18 +8,18 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.excellaproject.models.Form_C;
 
-public interface Form_CRepository extends CrudRepository <Form_C, Long> {
+public interface Form_CRepository extends CrudRepository <Form_C, Integer> {
 	List<Form_C> findAll();
 	
-	Optional<Form_C> findByUser_id(long id);
+	Optional<Form_C> findByUser_id(int id);
 	
 	@Query(value="SELECT * from forms_c WHERE user_id = ?1 AND form_level = 0", nativeQuery=true)
-    List<Form_C> findRejectedForms(long id);
+    List<Form_C> findRejectedForms(int id);
 	
 	@Query(value="SELECT * from forms_c WHERE user_id = ?1 AND form_level = 1", nativeQuery=true)
-    List<Form_C> findPendingForms(long id);
+    List<Form_C> findPendingForms(int id);
     
     @Query(value="SELECT * from forms_c WHERE user_id = ?1 AND form_level = 2", nativeQuery=true)
-    List<Form_C> findCompleteForms(long id);
+    List<Form_C> findCompleteForms(int id);
 }
