+++ ExcellaPracticeProject/ExcellaProject/src/main/java/com/excellaproject/repositories/Form_DRepository.java
@@ -8,17 +8,17 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.excellaproject.models.Form_D;
 
-public interface Form_DRepository extends CrudRepository <Form_D, Long> {
+public interface Form_DRepository extends CrudRepository <Form_D, Integer> {
 	List<Form_D> findAll();
 	
-	Optional<Form_D> findByUser_id(long id);
+	Optional<Form_D> findByUser_id(int id);
 	
 	@Query(value="SELECT * from forms_d WHERE user_id = ?1 AND form_level = 0", nativeQuery=true)
-    List<Form_D> findRejectedForms(long id);
+    List<Form_D> findRejectedForms(int id);
 	
 	@Query(value="SELECT * from forms_d WHERE user_id = ?1 AND form_level = 1", nativeQuery=true)
-    List<Form_D> findPendingForms(long id);
+    List<Form_D> findPendingForms(int id);
     
     @Query(value="SELECT * from forms_d WHERE user_id = ?1 AND form_level = 2", nativeQuery=true)
-    List<Form_D> findCompleteForms(long id);
+    List<Form_D> findCompleteForms(int id);
 }
